Python Files, CSV File, C File
The Python file runs with the CSV (that's where it gets its information about vehicles from). It also has a test program that is the 2nd Python file. I also made a similar calculator in C but it uses pointers and direct input for the vehicle information instead of a CSV.